Don't split on space
`Process` wants an array of strings to pass as parameters. While it definitely looks simpler to just pass in a single string and split it based on space, this can cause odd problems in weird cases (for example: spaces in paths).

This commit changes the simctl method to accept a variadic array of parameters, so that it doesn't have to do error-prone string splitting