Filter the list of simulators
This adds a filter field at the bottom of the sidebar, like the filter field in Xcode's project navigator.

Part of this change also moves the logic for retrieving and filtering the list of simulators to a `SimulatorsController` class, which handles loading, parsing, and filtering the simulators. That object then gets passed around and observed by various views.