Device state
This adds the concept of "device state", which corresponds to whether the simulator has booted or is shutdown (there are a number of possible states, but those are the main ones).

This state is kept fresh by polling `simctl` every 5 seconds for new information. Controls for non-booted simulators are disabled