Use CoreSimulator framework, when possible
CoreSimulator is a private framework we can use to be notified of changes to the simulators, without having to poll every 5 seconds. We bridge this functionality into the app from Objective-C by way of a custom Combine publisher, which makes it seamlessly fit in with the existing publishers.

We fall back to polling for information in the case that CoreSimulator cannot be loaded.